+++ example/macos/Flutter/GeneratedPluginRegistrant.swift
@@ -7,8 +7,10 @@ import Foundation
 
 import decrypt_pdf
 import file_picker
+import pdfx
 
 func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
   DecryptPdfPlugin.register(with: registry.registrar(forPlugin: "DecryptPdfPlugin"))
   FilePickerPlugin.register(with: registry.registrar(forPlugin: "FilePickerPlugin"))
+  PdfxPlugin.register(with: registry.registrar(forPlugin: "PdfxPlugin"))
 }