+++ example/windows/flutter/generated_plugin_registrant.cc
@@ -7,8 +7,11 @@
 #include "generated_plugin_registrant.h"
 
 #include <decrypt_pdf/decrypt_pdf_plugin_c_api.h>
+#include <pdfx/pdfx_plugin.h>
 
 void RegisterPlugins(flutter::PluginRegistry* registry) {
   DecryptPdfPluginCApiRegisterWithRegistrar(
       registry->GetRegistrarForPlugin("DecryptPdfPluginCApi"));
+  PdfxPluginRegisterWithRegistrar(
+      registry->GetRegistrarForPlugin("PdfxPlugin"));
 }