+++ lib/decrypt_pdf.dart
@@ -1,22 +1,8 @@
 import 'dart:async';
-import 'dart:io';
-
-import 'package:flutter/foundation.dart';
 
 import 'decrypt_pdf_platform_interface.dart';
-// Conditionally import Windows implementation
-import 'decrypt_pdf_windows.dart'
-    if (dart.library.io) 'decrypt_pdf_windows.dart';
 
 class DecryptPdf {
-  static void registerWith() {
-    if (kIsWeb) {
-      // Web registration happens in web file
-    } else if (Platform.isWindows) {
-      DecryptPdfWindows.registerWith();
-    }
-  }
-
   static Future<String?> getPlatformVersion() {
     return DecryptPdfPlatform.instance.getPlatformVersion();
   }
@@ -30,6 +16,17 @@ class DecryptPdf {
     required String password,
   }) {
     return DecryptPdfPlatform.instance.openPdf(
+      filePath: filePath,
+      password: password,
+    );
+  }
+
+  //get getPdfAsBase64
+  static Future<String?> getPdfAsBase64({
+    required String filePath,
+    required String password,
+  }) {
+    return DecryptPdfPlatform.instance.getPdfAsBase64(
       filePath: filePath,
       password: password,
     );