+++ lib/decrypt_pdf_method_channel.dart
@@ -24,7 +24,9 @@ class MethodChannelDecryptPdf extends DecryptPdfPlatform {
       });
       return result;
     } on PlatformException catch (e) {
-      print("Failed to check if PDF is protected: '${e.message}'.");
+      if (kDebugMode) {
+        print("Failed to check if PDF is protected: '${e.message}'.");
+      }
       rethrow;
     }
   }
@@ -41,7 +43,28 @@ class MethodChannelDecryptPdf extends DecryptPdfPlatform {
       );
       return decryptedFilePath;
     } on PlatformException catch (e) {
-      print("Failed to open PDF: '${e.message}'.");
+      if (kDebugMode) {
+        print("Failed to open PDF: '${e.message}'.");
+      }
+      rethrow;
+    }
+  }
+
+  @override
+  Future<String?> getPdfAsBase64({
+    required String filePath,
+    required String password,
+  }) async {
+    try {
+      final String? base64String = await methodChannel.invokeMethod(
+        'getPdfAsBase64',
+        {'filePath': filePath, 'password': password},
+      );
+      return base64String;
+    } on PlatformException catch (e) {
+      if (kDebugMode) {
+        print("Failed to get PDF as Base64: '${e.message}'.");
+      }
       rethrow;
     }
   }